+++ ticketgol/guardias/src/main/java/ticketgol/guardias/service/GuardiasService.java
@@ -3,9 +3,9 @@ package ticketgol.guardias.service;
 import org.springframework.stereotype.Service;
 import ticketgol.guardias.model.Guardias;
 import ticketgol.guardias.repository.GuardiasRepository;
+import ticketgol.guardias.exception.GuardiaNotFoundException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class GuardiasService {
@@ -19,8 +19,9 @@ public class GuardiasService {
         return repository.findAll();
     }
 
-    public Optional<Guardias> obtenerPorId(Long id) {
-        return repository.findById(id);
+    public Guardias obtenerPorId(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new GuardiaNotFoundException("El guardia con ID " + id + " no existe."));
     }
 
     public Guardias guardar(Guardias guardia) {
@@ -28,12 +29,13 @@ public class GuardiasService {
     }
 
     public void eliminar(Long id) {
+        if (!repository.existsById(id)) {
+            throw new GuardiaNotFoundException("No se puede eliminar. El guardia con ID " + id + " no existe.");
+        }
         repository.deleteById(id);
     }
 
-    // buscar guardias  con estadio
     public List<Guardias> obtenerPorEstadio(Long estadioId) {
         return repository.findByEstadioId(estadioId);
     }
 }
-
